pyproject.toml
[project]
name = "tiny-tg"
version = "0.1.0"
description = "Simple Telegram notification library"
authors = [{name = "rickhehe", email = "rick@rickhehe.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/rickhehe/tiny-tg"
Repository = "https://github.com/rickhehe/tiny-tg"

[project.scripts]
tg = "tiny_tg.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tiny_tg"]